Remove deprecated features; bump minimum version
The package list is now directly supplied to "apt" module instead of
using a loop and relying on implicit squash_actions feature.  The old
method will be removed in Ansible 2.11.

More information:
https://docs.ansible.com/ansible/2.7/porting_guides/porting_guide_2.7.html#using-a-loop-on-a-package-module-via-squash-actions

Using tests as filters is deprecated and will be removed in Ansible 2.9.
Thus, the pipe "|" used with version_compare is changed to "is".  The
test "version_compare" was renamed to "version" in Ansible 2.5, so the
minimum Ansible version is changed as well.

More information:
https://docs.ansible.com/ansible/2.7/dev_guide/testing/sanity/no-tests-as-filters.html
https://docs.ansible.com/ansible/2.7/user_guide/playbooks_tests.html#version-comparison